Set up key to track changes.
Th general rule is that all entries are “implemented”, because most of
them actually are. I’m doing that with query, so theres nothing to add
to the code itself. If the feature is not yet implemented (can be ”in
progress”, “future version”, or “removed”), each link is identified
with a class that will attach an icon to the list, helping identify the
status of that feature.

Then, while I’m actually adding the translated examples, I’m also
adding a “done” or a “revisit” class to the link, that will change
color form grey to blue/orange, to signify progress.